Rename package name of shodo.aio
